+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Modified
+
+- [TD-967] New endpoint created to retrieve the rule detail with the possible system params used to create an implementation 
+
 ## [2.12.0] 2019-01-29
 
 ### Added